Translate the buffered-texture cluster, and drop a fabricated decompressor
decompress_gfx_rom was answered on PC by a hand-written GBA-BIOS LZ77/RLE
decompressor taking (src, dest).  That was wrong twice over: the game's
routine is not a BIOS decompressor, and decompress_gfx_init calls it with a
single progress-struct argument.  Any live call would have read a bogus
header out of the progress struct and written through a leftover register.

platform/gfx_decompress.c now carries a translation of the real ARM routine
(asm/code_08000a00.s:353-502).  The format is two independent bit windows
refilled 32 bits at a time: window1 picks raw halfword vs run, window2 picks
a 1bpp run (two 4-bit colours, count in the high byte of the same halfword)
vs a 2bpp run (four colours, count in the low byte of the first selector).
`lim1` is a work budget, so a long texture spreads over frames.

The rest of the cluster is translated from assembly too, replacing stubs:
func_0800861c/08008628/08008658/0800869c (the cache) and func_08008608/
08008720 (expand one texture now).  func_0800869c is the reason the cache
entry embeds a CompressedData: it hands that back to the graphics-table
loader as a description of the same texture with the expansion already done.

Fixes two bugs in the loader translation from the previous commit's working
tree: the buffer was sized from CompressedGFX.count (compression units)
instead of .size (expanded halfwords), and update_texture_loader_task passed
decompress_gfx_init a limit of 0x7fffffff instead of the 0x1000 per-frame
budget the assembly uses, which would have stalled a frame on a long list.

No runtime change on PC, and none was expected: platform/gfx_data.c builds
every CompressedData with doubleCompressed = 0, because the asset pipeline
extracts graphics already decompressed.  So this path is unreachable in the
current build and could not be verified at runtime — it is verified against
the assembly only.  A 15-scene sweep against fd2d97bb is identical except
for one colour of frame-sampling noise in sick_beats.

This also retracts the reasoning in the previous working tree, which held
that engines drew flat fills because the texture loader was stubbed out.
They do not: the loader has nothing to decompress on PC.  Its only real job
here is to be a task that completes, so the run_func_after_task chain that
follows it fires — and the stand-in it replaces already did that.

+++ src/graphics_table.h
@@ -7,7 +7,10 @@
 
 struct GFXDecompressProgress {
     const u16 *data;
-    u32 size;
+    // Named "size" by the decompilation, but every caller passes a destination
+    // address (graphics_table.c computes info->dest + offset - ...), so it has
+    // to be pointer-width.
+    uintptr_t size;
     u32 count;
     u32 curwin1;
     u32 curwin2;
